# openapi.yaml
openapi: 3.0.3
info:
  title: Lifebuddy API
  description: |
    Habits, tasks, goals, and households. Most endpoints require a JWT from `POST /auth/login`.
    Use the token in the **Authorization** header as `Bearer <token>`.
  version: 1.0.0
servers:
  - url: https://your-api-id.execute-api.us-east-1.amazonaws.com
    description: Replace with your SAM-deployed API URL (e.g. from `sam deploy` output)
  - url: http://localhost:3001
    description: Local (e.g. SAM local start-api)
security:
  - bearerAuth: []
paths:
...
  /auth/signup:
    post:
      summary: Create account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    type: object
                    properties:
                      userId: { type: string }
                      email: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: User already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorMessage" }
  /auth/login:
    post:
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    type: object
                    properties:
                      userId: { type: string }
                      email: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorMessage" }
  /habits:
    get:
      summary: List habits
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Habit" }
    post:
      summary: Create habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                frequency: { type: string, enum: [daily, weekly, monthly, custom], default: daily }
                preferredTimeOfDay: { type: string, enum: [morning, afternoon, evening, any], default: any }
                dayOfWeek: { type: integer, minimum: 0, maximum: 6, description: "0-6 Sun-Sat, for weekly" }
                dayOfMonth: { type: integer, minimum: 1, maximum: 31, description: "For monthly" }
                reminderChannel: { type: string, enum: [none, in_app, email] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Habit" }
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /tasks:
    get:
      summary: List tasks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                dueDate: { type: string, format: date }
                durationMinutes: { type: integer }
                status: { type: string, enum: [pending, done, cancelled], default: pending }
                recurrence: { type: object, properties: { unit: { type: string, enum: [days, months] }, value: { type: integer } } }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /goals:
    get:
      summary: List goals
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Goal" }
    post:
      summary: Create goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                targetDate: { type: string, format: date }
                status: { type: string, enum: [active, completed, paused], default: active }
                linkedTaskIds: { type: array, items: { type: string }, default: [] }
                linkedHabitIds: { type: array, items: { type: string }, default: [] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Goal" }
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /households:
    get:
      summary: List my households (memberships)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HouseholdMember" }
    post:
      summary: Create household
      description: Creates a new household and adds you as owner. Use this to create the first household, then add members with POST /households/{householdId}/members.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, example: "Smith Family" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  household:
                    type: object
                    properties:
                      householdId: { type: string }
                      name: { type: string }
                      createdAt: { type: string }
                      updatedAt: { type: string }
                  membership:
                    $ref: "#/components/schemas/HouseholdMember"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token from POST /auth/login
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorMessage" }
...
  schemas:
    ErrorMessage:
      type: object
      properties:
        message: { type: string }
    Habit:
      type: object
      properties:
        habitId: { type: string }
        userId: { type: string }
        name: { type: string }
        description: { type: string }
        frequency: { type: string, enum: [daily, weekly, monthly, custom] }
        preferredTimeOfDay: { type: string, enum: [morning, afternoon, evening, any] }
        dayOfWeek: { type: integer }
        dayOfMonth: { type: integer }
        reminderChannel: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Task:
      type: object
      properties:
        taskId: { type: string }
        userId: { type: string }
        title: { type: string }
        description: { type: string }
        dueDate: { type: string, format: date }
        durationMinutes: { type: integer }
        recurrence: { type: object }
        status: { type: string, enum: [pending, done, cancelled] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Goal:
      type: object
      properties:
        goalId: { type: string }
        userId: { type: string }
        title: { type: string }
        description: { type: string }
        targetDate: { type: string, format: date }
        status: { type: string, enum: [active, completed, paused] }
        linkedTaskIds: { type: array, items: { type: string } }
        linkedHabitIds: { type: array, items: { type: string } }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    HouseholdMember:
      type: object
      properties:
        householdId: { type: string }
        userId: { type: string }
        role: { type: string, enum: [owner, member] }
        joinedAt: { type: string }
        householdName: { type: string }